PvP mode added

+++ src/TicTacToeUI/src/app/(components)/LoginPage.tsx
@@ -2,6 +2,7 @@
 import { useState } from "react";
 import { useRouter } from "next/navigation";
 import { useCookies } from "react-cookie";
+import Swal from "sweetalert2";
 
 export default function LoginPage() {
   const router = useRouter();
@@ -20,7 +21,11 @@ export default function LoginPage() {
     e.preventDefault();
 
     if (loginCredential.username === "" || loginCredential.password === "") {
-      window.alert("Please Fill in the Sign Up Form Completely!");
+      Swal.fire({
+        title: "Error!",
+        text: "Please Fill in the Login Form Completely!",
+        icon: "warning",
+      });
     } else {
       const res = await fetch("http://localhost:8080/database/login", {
         method: "POST",
@@ -33,44 +38,49 @@ export default function LoginPage() {
         },
       });
       const message = await res.json();
-      console.log(message);
 
       if (message.message === "Successfully Login!") {
-        window.alert(message.message);
+        Swal.fire({ title: message.message, icon: "success" }).then(
+          async (result) => {
+            if (result.isConfirmed) {
+              const response = await fetch("/api/loginApi", {
+                method: "POST",
+                headers: {
+                  "Content-Type": "application/json",
+                },
+                body: JSON.stringify({ username: loginCredential.username }),
+              });
 
-        const response = await fetch("/api/loginApi", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ username: loginCredential.username }),
-        });
+              const res = await response.json();
+              const token = res.token;
 
-        const res = await response.json();
-        const token = res.token;
+              const get = await fetch(
+                "http://localhost:8080/database/findemail",
+                {
+                  method: "POST",
+                  headers: { "Content-Type": "application/json" },
+                  body: JSON.stringify({ username: loginCredential.username }),
+                }
+              );
 
-        const get = await fetch("http://localhost:8080/database/findemail", {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ username: loginCredential.username }),
-        });
-        console.log(get);
-        const getString = await get.json();
-        const email = getString.email;
+              const getString = await get.json();
+              const email = getString.email;
 
-        window.localStorage.setItem("token", token);
-        window.localStorage.setItem("username", loginCredential.username);
-        window.localStorage.setItem("email", email);
-
-        router.push("/home");
+              window.localStorage.setItem("token", token);
+              window.localStorage.setItem("username", loginCredential.username);
+              window.localStorage.setItem("email", email);
+              router.push("/home");
+            }
+          }
+        );
       } else if (message.message === "Incorrect Password!") {
-        window.alert(message.message);
+        Swal.fire({ title: "Incorrect Password!", icon: "error" });
         setCredential((prev) => ({
           ...prev,
           password: "",
         }));
       } else {
-        window.alert(message.message);
+        Swal.fire({ title: message.message, icon: "error" });
         setCredential({
           username: "",
           password: "",
@@ -89,7 +99,7 @@ export default function LoginPage() {
           <h1 className="text-2xl font-bold">Login</h1>
           <div className="flex flex-col gap-4">
             <div className="flex gap-6 items-center">
-              <h1 className="w-[6rem] font-bold">Username : </h1>
+              <h1 className="w-[7rem] font-bold">Username : </h1>
               <input
                 className="rounded-xl p-[0.4rem_0.8rem] shadow-xl"
                 placeholder="Username"
@@ -100,7 +110,7 @@ export default function LoginPage() {
               />
             </div>
             <div className="flex gap-6 items-center">
-              <h1 className="w-[6rem] font-bold">Password : </h1>
+              <h1 className="w-[7rem] font-bold">Password : </h1>
               <input
                 className="rounded-xl p-[0.4rem_0.8rem] shadow-xl"
                 placeholder="Password"
